+++ src/index.js
@@ -1,4 +1,4 @@
-export { svgIcone, svgIconeAsync, resolverBruto, precarregar, listarIcones, temIcone, buscarIcones, listarCategorias, categoriaDoIcone, descricaoDoIcone, iconesPorCategoria, versaoDoIcone, iconesPorVersao, listarVersoes } from './core.js'
+export { svgIcone, svgIconeAsync, resolverBruto, precarregar, listarIcones, temIcone, buscarIcones, listarCategorias, categoriaDoIcone, descricaoDoIcone, iconesPorCategoria, versaoDoIcone, iconesPorVersao, listarVersoes, CONJUNTO_PADRAO, conjuntoDoIcone, listarConjuntos, iconesPorConjunto, mesmoConjunto } from './core.js'
 export { default as SvgIcone } from './SvgIcone.js'
 import { svgIcone } from './core.js'
 export default svgIcone